Feat - Bootstrap application
The project is divided into two parts:

The first part is to publish in a kafka the results of the French presidential election 2017.
The data are stored in a csv file. We parse each line individually and we serialize to json format in kafka.

The second part consists of retrieving the kafka data and printing it in a console.

Miscellaneous:

For the deployment of the kafka, a dockerfile was created.